chore: pin package versions to patch level
ESLint plugins often introduce new features in minor releases, and
occasionally enable backward-incompatible rules in minor releases. I
would rather be forced to refer to changelogs for non-buglevel fixes to
be sure the lint output is stable across installations and to force me
to consider new rules as they're introduced.